Move saneyaml.py and tests from licensedcode to commoncode #831
