+++ src/commoncode/saneyaml.py
